# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Open Brain MCP API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Thought {
  id: number;
  content: string;
  category: string;
  tags: string[];
  pinned: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateThoughtInput {
  content: string;
  category?: string;
  tags?: string[];
  pinned?: boolean;
}

export interface UpdateThoughtInput {
  content?: string;
  category?: string;
  tags?: string[];
  pinned?: boolean;
}

export interface ErrorResponse {
  error: string;
}

export type ListThoughtsParams = {
  /**
   * Filter by category
   */
  category?: string;
  /**
   * Search by text
   */
  search?: string;
};