# lib/api-zod/src/generated/types/thought.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Open Brain MCP API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Thought {
  id: number;
  content: string;
  category: string;
  tags: string[];
  pinned: boolean;
  createdAt: Date;
  updatedAt: Date;
}